% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.R
\name{chi2}
\alias{chi2}
\title{Compute Chi2 goodness of fit statistic between two p-curves}
\usage{
chi2(v1, v2, n)
}
\arguments{
\item{v1}{First vector of probabilities (original)}

\item{v2}{Second vector of probabilities (fitted values)}

\item{n}{Total sample size}
}
\description{
Calculates Chi squared statistic between two vectors
}
