% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi2.R
\name{chi2_old}
\alias{chi2_old}
\title{Compute Chi2 goodness of fit statistic between two p-curves}
\usage{
chi2_old(observed, expected, n)
}
\arguments{
\item{observed}{First vector of probabilities (original)}

\item{expected}{Second vector of probabilities (fitted values), can also be a matrix with each row being a different expected}

\item{n}{Total sample size}
}
\description{
Calculates Chi squared statistic between two vectors
}
