% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cumvar}
\alias{cumvar}
\title{Compute cumulative variance}
\usage{
cumvar(x)
}
\arguments{
\item{x}{Vector containing values for variance computation}
}
\description{
Iteratively compute variance of 1:k, 1:k+1, 1:k+2 datapoints in a vector
}
