% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optionalStopping.R
\name{optionalStopping}
\alias{optionalStopping}
\title{Run optional stopping and return p-curve}
\usage{
optionalStopping(
  simres,
  ES,
  prop_Hacker,
  prop_H1,
  nmin,
  nmax,
  stepsize,
  d,
  het,
  alpha = 0.05
)
}
\arguments{
\item{simres}{Simulation result created in sim_optionalStopping.R (array: nMax x iter x ncol(ES))}

\item{ES}{Corresponding effect size object (expand.grid(ds, het))}

\item{prop_Hacker}{Proportion of p-hackers in the population}

\item{prop_H1}{Proportion of H1 in the population}

\item{nmin}{Minimum sample size per group}

\item{nmax}{Maximum sample size per group}

\item{stepsize}{Every how many participants do you check?}

\item{d}{Effect size}

\item{het}{Heterogeneity}

\item{alpha}{Significance level}
}
\description{
Run optional stopping and return p-curve
}
