% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse_old}
\alias{rmse_old}
\title{Compute RMSE between two p-curves}
\usage{
rmse_old(observed, expected)
}
\arguments{
\item{observed}{First vector (original)}

\item{expected}{Second vector (fitted values)}
}
\description{
Calculates square root of mean of squared differences between two vectors
}
